+++ CHANGELOG.md
@@ -14,6 +14,7 @@ the Facebook Faiss team.  Feel free to add entries here if you submit a PR.
 - Added FlatDistanceComputer for all FlatCodes indexes
 - Support for fast accumulation of 4-bit LSQ and RQ
 - Added product additive quantization
+- Support per-query search parameters for many indexes + filtering by ids
 
 ## [1.7.2] - 2021-12-15
 ### Added